API Reference

CHMMAIRRa.detect_chimeras — Method

detect_chimeras(V_db::Tuple{Vector{String}, Vector{String}}, assignments::DataFrame;
                J_db::Union{Tuple{Vector{String}, Vector{String}}, Nothing} = nothing,
                p_threshold::Float64 = 0.95,
                min_DFR::Int = 1,
                HMM_parameters::Union{Dict, Nothing} = nothing,
                receptor::String = "IG",

                chimeric_alignments::Bool = false,
                recombfreqplot::Bool = false,
                detailed::Bool = false,
                subsample::Union{Int, Nothing} = nothing,
                deduplicate::Bool = false,

                count_chimeric_segments::Bool = false,
                trim::Int = 5,
                seed::Int = 123,

                mafft::Union{String, Nothing} = nothing,
                align_database::Bool = false,

                quiet::Bool = false,
                disable_internal_dedup::Bool = false
                )::NamedTuple{(:out, :chimeric_alignments, :recombfreqplot),
                            Tuple{DataFrame,
                            Union{Nothing, Tuple{Vector{String}, Vector{String}}},
                            Any}
                            }
                }

Run CHMMAIRRa on parsed arguments, returning output as a Tuple containing relevant outputs. Output tuple fields:

out: the input assignments dataframe with chimerism probabilities.

chimericalignments: names and sequences for alignments showing which reference sequences recombined at which positions to form detected chimeras. Set to nothing if chimericalignments is false.

recombfreqplot: barplot of most frequent recombinations, normalized by their expected frequencies. Set to nothing if recombfreqplot is false.

Arguments:

V_db::Tuple{Vector{String}, Vector{String}}: Tuple containing (reference names, reference sequences) for the V database.

assignments::DataFrame: VDJ assignments, in MiAIRR format.

Optional arguments:

J_db::Union{Tuple{Vector{String}, Vector{String}}, Nothing}: Optional tuple containing (reference names, reference sequences) for the J database. If provided, will run on both V and J and the sequence will be chimeric if either V or J is chimeric.

p_threshold::Float64: Posterior threshold of switching templates.

min_DFR::Int: Minimum differences from reference (DFR) threshold to consider a recombination event.

HMM_parameters::Union{Dict, Nothing}: If set, use HMM parameters in this dictionary. Overrides parameter presets from receptor.

receptor::String: Which adaptive immune receptor to run on. Options: IG, TCR.

chimeric_alignments::Bool: Whether to generate and return chimeric sequences and the germline alignments they matched to.

recombfreqplot::Bool: Whether to generate and return a plot of chimerism per recombination event.

detailed::Bool: Whether to include details about the chimeras. Calculate viterbi path for chimeric sequences to add startingpoint, recombinations, recombinationsdegapped, and maxlog_prob columns.

subsample::Union{Int, Nothing}: Randomly subsample this number of unique sequences to run chimera detection on.

deduplicate::Bool: Whether to deduplicate input by vsequencealignment column. Note that if subsample is given, subsampling is performed after deduplication.

count_chimeric_segments::Bool: If set, count the number of times chimeric segments appear in nonchimeric sequences. Adds chimerasegmentcounts column.

trim::Int: How many nucleotides to trim off the inner ends of the chimeric segments when searching for matches in nonchimeric sequences.

seed::Int: Seed to use for random subsampling.

mafft::Union{String, Nothing}: Path to MAFFT binary. If not provided, will try to find mafft in system PATH.

align_database::Bool: Whether to align database sequences.

quiet::Bool: If set, hides non-error messages.

disable_internal_dedup::Bool: If set, does not deduplicate sequences internally. This was added for timing benchmark purposes and has no effect on the output.

CHMMAIRRa.detect_chimeras — Method

detect_chimeras(refnames::Vector{String}, refseqs::Vector{String}, assignments::DataFrame, gene::Char;
                p_threshold::Float64 = 0.95,
                min_DFR::Int = 1,
                HMM_parameters::Union{Dict, Nothing} = nothing,
                receptor::String = "IG",

                chimeric_alignments::Bool = false,
                recombfreqplot::Bool = false,
                detailed::Bool = false,
                subsample::Union{Int, Nothing} = nothing,
                deduplicate::Bool = false,

                count_chimeric_segments::Bool = false,
                trim::Int = 5,
                seed::Int = 123,

                mafft::Union{String, Nothing} = nothing,
                align_database::Bool = false,

                quiet::Bool = false,
                disable_internal_dedup::Bool = false
                )::NamedTuple{(:out, :chimeric_alignments, :recombfreqplot),
                            Tuple{DataFrame,
                            Union{Nothing, Tuple{Vector{String}, Vector{String}}},
                            Any}
                            }
                }

Run CHMMAIRRa on parsed arguments, returning output as a Tuple containing relevant outputs. Output tuple fields:

out: the input assignments dataframe with chimerism probabilities.

chimericalignments: names and sequences for alignments showing which reference sequences recombined at which positions to form detected chimeras. Set to nothing if chimericalignments is false.

recombfreqplot: barplot of most frequent recombinations, normalized by their expected frequencies. Set to nothing if recombfreqplot is false.

Arguments:

refnames::Vector{String}: Names of the reference sequences.

refseqs::Vector{String}: Sequences of the reference sequences.

assignments::DataFrame: VDJ assignments, in MiAIRR format.

gene::Char: Which gene to run on. Options: 'v', 'j'.

Optional arguments:

p_threshold::Float64: Posterior threshold of switching templates.

min_DFR::Int: Minimum differences from reference (DFR) threshold to consider a recombination event.

HMM_parameters::Union{Dict, Nothing}: If set, use HMM parameters in this dictionary. Overrides parameter presets from receptor.

receptor::String: Which adaptive immune receptor to run on. Options: IG, TCR.

chimeric_alignments::Bool: Whether to generate and return chimeric sequences and the germline alignments they matched to.

recombfreqplot::Bool: Whether to generate and return a plot of chimerism per recombination event.

detailed::Bool: Whether to include details about the chimeras. Calculate viterbi path for chimeric sequences to add startingpoint, recombinations, recombinationsdegapped, and maxlog_prob columns.

subsample::Union{Int, Nothing}: Randomly subsample this number of unique sequences to run chimera detection on.

deduplicate::Bool: Whether to deduplicate input by vsequencealignment column. Note that if subsample is given, subsampling is performed after deduplication.

count_chimeric_segments::Bool: If set, count the number of times chimeric segments appear in nonchimeric sequences. Adds chimerasegmentcounts column.

trim::Int: How many nucleotides to trim off the inner ends of the chimeric segments when searching for matches in nonchimeric sequences.

seed::Int: Seed to use for random subsampling.

mafft::Union{String, Nothing}: Path to MAFFT binary. If not provided, will try to find mafft in system PATH.

align_database::Bool: Whether to align database sequences.

quiet::Bool: If set, hides non-error messages.

disable_internal_dedup::Bool: If set, does not deduplicate sequences internally. This was added for timing benchmark purposes and has no effect on the output.

CHMMAIRRa.detect_chimeras_from_files — Method

detect_chimeras_from_files(V_fasta::String, assignments::String, out::String;

        receptor::String = "IG",
        J_fasta::Union{String, Nothing} = nothing,

        non_chimeric_MiAIRR::Union{String, Nothing} = nothing,
        chimeric_MiAIRR::Union{String, Nothing} = nothing,
        chimeric_alignments::Union{String, Nothing} = nothing,
        recombfreqplot::Union{String, Nothing} = nothing,
        detailed::Bool = false,
        subsample::Union{Int, Nothing} = nothing,
        deduplicate::Bool = false,
        chunk_size::Union{Int, Nothing} = nothing,
       
        p_threshold::Float64 = 0.95,
        min_DFR::Int = 1,
        HMM_parameters::Union{String, Nothing} = nothing,
        mafft::Union{String, Nothing} = nothing,
        align_database::Bool = false,
        count_chimeric_segments::Bool = false,
        trim::Int = 5,
        seed::Int = 123,

        quiet::Bool = false,
        disable_internal_dedup::Bool = false
        )

Run CHMMAIRRa from input files to writing of output files, taking care of I/O.

Arguments:

V_fasta::String: V database in fasta format.

assignments::String: VDJ assignments, in MiAIRR format.

out::String: Write CHMMAIRRa output.

Optional arguments:

receptor::String: Which adaptive immune receptor to run on. Options: IG, TCR.

J_fasta::Union{String, Nothing}: J database in fasta format. If provided, will run on both V and J and the sequence will be chimeric if either V or J is chimeric.

non_chimeric_MiAIRR::Union{String, Nothing}: Write assignments of sequences determined to be non-chimeric.

chimeric_MiAIRR::Union{String, Nothing}: Write assignments of sequences determined to be chimeric.

chimeric_alignments::Union{String, Nothing}: Write fasta file containing chimeric sequences and the germline alignments they matched to.

recombfreqplot::Union{String, Nothing}: Generate a plot of chimerism per recombination event.

detailed::Bool: Include details about the chimeras. Calculate viterbi path for chimeric sequences to add startingpoint, recombinations, recombinationsdegapped, and maxlog_prob columns.

subsample::Union{Int, Nothing}: Randomly subsample this number of unique sequences to run chimera detection on.

deduplicate::Bool: Deduplicate input by vsequencealignment column. Note that if subsample is given, subsampling is performed after deduplication.

chunk_size::Union{Int, Nothing}: Read in chunk-size lines of input at a time, write output for that chunk, and move to the next chunk.

p_threshold::Float64: Posterior threshold of switching templates.

min_DFR::Int: Minimum differences from reference (DFR) threshold to consider a recombination event.

HMM_parameters::Union{String, Nothing}: If set, use HMM parameters in this file. Overrides parameter presets from –receptor. See src/IGparameters.tsv or src/TCRparameters.tsv for formatting.

mafft::Union{String, Nothing}: Path to MAFFT binary. If not provided, will try to find mafft in system PATH.

align_database::Bool: Whether to align database sequences.

count_chimeric_segments::Bool: If set, count the number of times chimeric segments appear in nonchimeric sequences. Adds chimerasegmentcounts column.

trim::Int: How many nucleotides to trim off the inner ends of the chimeric segments when searching for matches in nonchimeric sequences.

seed::Int: Seed to use for random subsampling.

quiet::Bool: If set, hides non-error messages.

disable_internal_dedup::Bool: If set, does not deduplicate sequences internally. This was added for timing benchmark purposes and has no effect on the output.

CHMMAIRRa.julia_main — Method

julia_main()::Cint

Entry point C call for main function for Julia